+++ src/publisher/react-integration/utils/getDisplayName.ts
@@ -1,7 +1,7 @@
 const cachedDisplayNames = new WeakMap<any, string>();
-let anonymousSeedId = 1;
+const usedDisplayNames = new Map<string, number>();
 
-export function getDisplayName(type: any): string {
+export function getDisplayName(type: any, kind = ""): string {
   const displayNameFromCache = cachedDisplayNames.get(type);
   if (typeof displayNameFromCache === "string") {
     return displayNameFromCache;
@@ -21,7 +21,15 @@ export function getDisplayName(type: any): string {
   }
 
   if (!displayName) {
-    displayName = "Anonymous" + String(anonymousSeedId++);
+    displayName = "Anonymous" + kind;
+  }
+
+  if (usedDisplayNames.has(displayName)) {
+    const num = usedDisplayNames.get(displayName) || 2;
+    usedDisplayNames.set(displayName, num + 1);
+    displayName += `\`${String(num)}`;
+  } else {
+    usedDisplayNames.set(displayName, 2);
   }
 
   cachedDisplayNames.set(type, displayName);